move common tables into feature
