+++ cargo-shuttle/Cargo.toml
@@ -10,7 +10,7 @@ default-run = "shuttle"
 
 [dependencies]
 shuttle-api-client = { workspace = true, default-features = true }
-shuttle-common = { workspace = true, features = ["models"] }
+shuttle-common = { workspace = true, features = ["models", "tables"] }
 shuttle-service = { workspace = true, features = ["builder"] }
 
 anyhow = { workspace = true }
