+++ common/src/models/user.rs
@@ -1,9 +1,11 @@
-use std::fmt::Write;
-
 use chrono::{DateTime, Utc};
-use crossterm::style::Stylize;
 use serde::{Deserialize, Serialize};
 use strum::{Display, EnumString};
+
+#[cfg(feature = "display")]
+use crossterm::style::Stylize;
+#[cfg(feature = "display")]
+use std::fmt::Write;
 
 /// In normal cases, a string with the format `user_<ULID>`.
 /// This is a soft rule and the string can be something different.
@@ -21,6 +23,7 @@ pub struct UserResponse {
 }
 
 impl UserResponse {
+    #[cfg(feature = "display")]
     pub fn to_string_colored(&self) -> String {
         let mut s = String::new();
         writeln!(&mut s, "{}", "Account info:".bold()).unwrap();
